fix: Make shell exit if 'cd ..' is used for deleted parent dir
Up until now there was a check for it chdir fails within cd.
However this function will still return 0 even if the parent folder
(the one we want to go to) has been deleted (while using "..")
This version implements a function that changes ".." with the actual
path of the parent directory so it will actually detect that the
path is not valid and exit the program.